requirements.txt
# OpenAI + základní komunikace
openai
python-dotenv

# Vektorová databáze a embeddingy
chromadb
PyMuPDF        # pro čtení PDF souborů
tiktoken       # pokud používáš OpenAI pro tokenizaci

# LangChain
langchain
langchain-openai
langchain-community
langchain-chroma

# LlamaIndex
llama-index
llama-index-core
llama-index-embeddings-openai
llama-index-vector-stores-chroma
llama-index-llms-openai

# Semantic Kernel
semantic-kernel

# Frontendy (UI)
streamlit
flask
gradio

# Pomocné knihovny
uuid
requests

##pip install -r requirements.txt
